+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Changed
 - Fixed Diancie-Mega bankable flag being true
 - Fixed detail appearance of grid showing empty moves
+- Fixed missing icon for RNGs
 
 
 ## 0.4.2 - 2022-03-19